[Refactor] Create a default relationship document validator
The validator factory can now construct a default relationship
document validator - i.e. one that checks the document conforms
to the JSON API spec. This default one knows nothing of the
business logic for the document, e.g. what resource type to accept
or whether the relationship should be a has-one or has-many
relationship.

Refer to issue #23