[Feature] Add a generic validator provider
The generic validator provider is able to validate any request
content is valid according to the JSON API spec. The generic
implementation uses the new generic validators for a resource
document and relationship document.

As part of this, the validator provider interface was modified
so that it receives the values that are known from a request
URL. This means that a validator provider can now be used for
multiple different resource types if needed.

Refer to issue #23