[Refactor] Update the API factory interface to not receive config
...instead config should be injected into the actual API factory
instance. A `configure` method has been added to the API factory
class for this purpose.

The factory itself only now receives the namespace of the API to
be built plus the HTTP scheme and host to use (if needed). This
means the structure of configuration arrays is entirely at the
discretion of the factory instance itself.