[Feature] Reject related resource if type is unknown
If the resource type of a related resource is not recognised, i.e.
is not known to the store, then it is rejected with a specific
validation message.

This fixes a previous bug whereby submitting an unknown resource
type caused the store to throw a runtime exception because it
could not locate an adapter for the specified type.

It is also an improvement because it provides a clear validation
error message to the client that the resource type itself is not
known.

Fixes issue #27