[Bugfix] Ensure correct validation of resource identifier
Fixes #35, where submitting a null id in a relationship resource
identifier resulted in an internal server error when the
identifier was checked against the store.

Also ensured that the id member is always a string. Previously
integers were allowed, however the spec says the id member must
be a string.