[Bugfix] Use the content-length value when checking for content
Have observed some scenarios where a browser is sending an empty
string for the HTTP `Content-Length` header. Before we were
checking for the existence of the header when determining if a
message contained body. So in these scenarios, the content
was being decoded and a JSON decoding exception thrown.

Changed so that if there is a Content-Length header, it is now
checked to ensure it is a non-zero value. This means if a HTTP
request indicates that its content length is 0, it will now be
determined not to have message body.